requirements.txt
absl-py==2.1.0
accelerate==0.34.0
antlr4-python3-runtime==4.9.3
anyio==4.4.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
attrs==24.2.0
babel==2.16.0
beautifulsoup4==4.12.3
bitsandbytes==0.43.3
bleach==6.1.0
box2d-py==2.3.5
certifi==2024.8.30
cffi==1.17.0
charset-normalizer==3.3.2
cloudpickle==3.0.0
comm==0.2.2
contourpy==1.3.0
cycler==0.12.1
debugpy==1.8.5
decorator==5.1.1
defusedxml==0.7.1
einops==0.8.0
etils==1.9.4
executing==2.1.0
Farama-Notifications==0.0.4
fastjsonschema==2.20.0
filelock==3.13.1
flash-attn==2.6.3
fonttools==4.53.1
fqdn==1.5.1
fsspec==2024.2.0
glfw==2.7.0
gymnasium==0.29.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
huggingface-hub==0.24.6
hydra-core==1.3.2
idna==3.8
imageio==2.35.1
imageio-ffmpeg==0.5.1
importlib_resources==6.4.4
ipykernel==6.29.5
ipython==8.27.0
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.3
json5==0.9.25
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
kiwisolver==1.4.5
kornia==0.7.3
kornia_rs==0.1.5
MarkupSafe==2.1.5
matplotlib==3.9.2
matplotlib-inline==0.1.7
meshcat==0.3.2
mistune==3.0.2
mpmath==1.3.0
mujoco==3.2.2
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.2.1
notebook==7.2.2
notebook_shim==0.2.4
numpy==2.1.0
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cudnn-cu11==9.1.0.70
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.3.0.86
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusparse-cu11==11.7.5.86
nvidia-nccl-cu11==2.20.5
nvidia-nvtx-cu11==11.8.86
omegaconf==2.3.0
opencv-python==4.10.0.84
overrides==7.7.0
packaging==24.1
pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.2.2
prometheus_client==0.20.0
prompt_toolkit==3.0.47
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pybullet==3.2.6
pycparser==2.22
pygame==2.6.0
Pygments==2.18.0
pyngrok==7.2.0
PyOpenGL==3.1.7
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-json-logger==2.0.7
pytz==2024.1
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.35.1
regex==2024.7.24
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.20.0
safetensors==0.4.4
scipy==1.14.1
seaborn==0.13.2
Send2Trash==1.8.3
setuptools==70.0.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
stack-data==0.6.3
swig==4.2.1
sympy==1.12
terminado==0.18.1
timm==0.9.10
tinycss2==1.3.0
tokenizers==0.19.1
torch==2.4.0+cu118
torchaudio==2.4.0+cu118
torchvision==0.19.0+cu118
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
transformers==4.44.2
transforms3d==0.4.2
triton==3.0.0
types-python-dateutil==2.9.0.20240821
typing_extensions==4.9.0
tzdata==2024.1
u-msgpack-python==2.8.0
uri-template==1.3.0
urllib3==2.2.2
# Editable Git install with no remote (VLA==0.0.1)
-e /home/vishwak/Desktop/projects/Robotics/VLA
wcwidth==0.2.13
webcolors==24.8.0
webencodings==0.5.1
websocket-client==1.8.0
wheel==0.44.0
zipp==3.20.1
